Initial commit. Modify the bevy breakout example
Add ability to shoot extra balls. They aren't setup for collision yet
Add ammo field to the Scoreboard, and increment as balls are shot. Plan is to have set ammo, and you can only shoot that many balls
Copied the necessary assets from the Bevy repository
Also set the Window width and height
Change colors of some text

Goal is to make it so shot balls can break the bricks