Update to be similar to newest version in bevy main branch
This means updating the bevy crate version to 0.11.0-dev, to do this
instead of explicitly setting the version, I gave the path to the Bevy repo and main branch.

There were lots of big updates since I last worked on this, because Bevy updated from 0.9 to 0.10
Furthermore, work is being done on Bevy 0.11.0, so I had to update lots of stuff. I did this by doing a diff
of my main.rs file and the breakout.rs file in the Bevy repo.
The one thing I didn't get working yet that was working before was resizing the Window. From bevy 0.9 to 0.10,
the way Windows are treated changed. For now, it is commented out